requirements.txt
numpy>=1.19.5
pandas>=1.1.5
matplotlib>=3.3.4
seaborn>=0.11.1
scikit-learn>=0.24.1
opencv-python>=4.5.1.48
Pillow>=8.1.0
tensorflow>=2.8.0
keras>=2.8.0
efficientnet>=1.1.1
fastapi
uvicorn
pydantic_settings